DiscordBot 2.0/requirements.txt
discord.py
flask
aiosqlite
python-dotenv
scikit-learn
pillow
yt-dlp
discord.py
yt-dlp
PyNaCl
spotipy
requests